Move all class definition utility widgets to OpenLIFULib (#231)
- This will make it easier to use those patterns elsewhere, such as in
  the users management table.
